% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{du_add_margin}
\alias{du_add_margin}
\title{Add margins considering all numeric values in a data frame}
\usage{
du_add_margin(
  df,
  margin = "both",
  verbose = TRUE,
  ignore_columns = NULL,
  enable_total_margin = TRUE,
  direction_total_margin = "column",
  fun.aggregate = sum,
  ...
)
}
\arguments{
\item{df}{data frame to add margin to.}

\item{margin}{valid values are "column", "row", "both". Append margins after
column, row or both.}

\item{verbose}{show or hide messages}

\item{enable_total_margin}{default to \code{TRUE}, whether to calcualte total
margin when \code{margin = "both"}}

\item{direction_total_margin, }{default to "column", direction for calculating
total margin, please read more in details.}

\item{fun.aggregate}{function used to calculate margin, default to sum,
can be one of min, max, median, sd, sum, mean or any other custom
functions which can reduce a column/row
automatically enabled.}

\item{...}{other arguments passed to  aggregate function}

\item{ingnoreCols}{numeric columns to ignore when adding margin,
can use column names or column indexes.}
}
\value{
A data frame with extra an extra
column/row as the total for that column/row
}
\description{
Adding margins to column, row or both, non-numeric columns are ignored.Use
argument \code{ingnoreCols} to skip numeric columns.
}
\details{
when \code{direction} is set to "both", total margin on bottom right of the
data frame may not make sense when using mean, sd or similar functions as
aggregating fucntion.

To  specify direction of calculation:
\code{direction_total_margin = "column"} to calculate by columns(horizontally),
\code{direction_total_margin = "row"} to calculate by rows(vertically)
}
\examples{
du_add_margin(iris)
du_add_margin(iris,verbose = FALSE)
du_add_margin(iris,fun.aggregate=mean,na.rm = TRUE)
# disable calculation for total margin
du_add_margin(
  iris,
  enable_total_margin = FALSE,
  fun.aggregate=sd,
  na.rm = TRUE
)
}
