% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table.R
\name{du_display_data_shape}
\alias{du_display_data_shape}
\title{display cells in a data frame that doesn't contain missing value}
\usage{
du_display_data_shape(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a graph that display the shape of data
}
\description{
display cells in a data frame that doesn't contain missing value
}
