% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{du_log_output}
\alias{du_log_output}
\title{Capture formatted output of any variable, which can be used for logging}
\usage{
du_log_output(x)
}
\arguments{
\item{x}{a R object}
}
\value{
A character vector with rows separated by \code{\\n}
}
\description{
Capture formatted output of any variable, which can be used for logging
}
\examples{
\dontrun{
cat(du_log_output(head(iris)))
cat(du_log_output(head(tibble::as_tibble(iris))))
cat(du_log_output(summary(lm(Sepal.Width ~ Sepal.Length,data = iris))))
}
}
