% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table.R
\name{du_read_excel_ranges}
\alias{du_read_excel_ranges}
\title{read multiple ranges from the same sheet in an excel file}
\usage{
du_read_excel_ranges(path, sheet = NULL, range, guess = TRUE)
}
\arguments{
\item{path}{file of excel file}

\item{sheet}{the sheet name or number to be read. If NULL, read the first sheet}

\item{range}{range(s) in the given sheet, can be a unnamed/named vector or list.
If named, names will be used as the name in the list of data frames}

\item{guess}{whether to guess column type, if FALSE, read all columns as text}
}
\value{
a list of data frames
}
\description{
read multiple ranges from the same sheet in an excel file
}
\examples{
\dontrun{
file_path <- system.file("examples/excel.xlsx",package = "dfutils")
range1 <- c("A1:E5","H7:L11")
range2 <- list(tabl1="A1:E5",table2="H7:L11")
du_read_excel_ranges(file_path,range = range1)
du_read_excel_ranges(file_path,range = range2,guess = FALSE)
}

}
