% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table.R
\name{du_split_df}
\alias{du_split_df}
\title{split a large table into smaller tables if there are blank columns or rows}
\usage{
du_split_df(df, complexity = 1)
}
\arguments{
\item{df}{a data frame}

\item{complexity}{default to 1,number of times to split a data frame by row and column.
If number increase by one, one more split(by both row and column) will be performed}
}
\value{
a list of data frames
}
\description{
split a large table into smaller tables if there are blank columns or rows
}
\details{
1. if you still see entire rows or columns missing. Please increase complexity
2. The list of data frames is roughly ordered from left to right and top to
bottom in the original data frame
}
\examples{
mydf <- as.data.frame(matrix(1:100,nrow = 10))
mydf[5,] <- NA
mydf[,7] <- NA
mydf[,8 ] <- NA
mydf[1,c(1,2,3)] <- NA
du_split_df(mydf)

}
