% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table.R
\name{du_split_direction}
\alias{du_split_direction}
\title{split a large table in one direction if there are blank columns or rows}
\usage{
du_split_direction(df, direction = "col", split_index = FALSE)
}
\arguments{
\item{df}{a data frame}

\item{direction}{"col" or "row", the direction to split the data frame}

\item{split_index}{whether to return the index where the split is made}
}
\value{
If split_index is \code{TRUE}, returns a list of data frame.
If split_index is \code{FALSE}, returns a list of two elements, the first element
\code{df} contains a list of split data frames, the second element \code{split_index}
is the index where the split is performed.
}
\description{
split a large table in one direction if there are blank columns or rows
}
