% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatrame_io.R
\name{du_write_excel}
\alias{du_write_excel}
\title{write a list of data frames to excel with each data frame in a tab}
\usage{
du_write_excel(x, path, add_index = FALSE, overwrite = FALSE)
}
\arguments{
\item{x}{a list of data frames}

\item{path}{the path to write}

\item{add_index}{whether to add index number in sheet name, if \code{TRUE}, tabs
are named such as 1_sheetName1, 2_sheetName2 ...}

\item{overwrite}{whether to overwrite existing file}
}
\value{
NULL, files written to an excel file.
}
\description{
Element names will be used as sheet name.
}
\examples{
\dontrun{
dflist <- list(iris = head(iris),mtcars = head(mtcars))
du_write_excel(dflist,"test.xlsx")
du_write_excel(dflist,"test.xlsx",add_index = TRUE, overwrite = TRUE)
}

}
