% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table.R
\name{make_df_index}
\alias{make_df_index}
\title{returns the consecutive row/column numbers that are not NA in a list}
\usage{
make_df_index(v)
}
\arguments{
\item{v}{a boolean vector(only TRUE/FALSE)}
}
\value{
A list of consecutive row/column numbers as different elements
}
\description{
Only accept a boolean vector(only TRUE or FALSE). Otherwise, this
function doesn't make any sense.
}
