% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_api.R
\name{safely_get_geocode}
\alias{safely_get_geocode}
\title{A warpper on \code{googleway::google_geocode()}: get geocode from google safely}
\usage{
safely_get_geocode(address, apikey)
}
\arguments{
\item{address}{address, vector length of one}

\item{apikey}{google's apikey}
}
\value{
a list of google's geocode service result
}
\description{
A warpper on \code{googleway::google_geocode()}: get geocode from google safely
}
