% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table.R
\name{vec_rle}
\alias{vec_rle}
\title{get rle as a named vector}
\usage{
vec_rle(v)
}
\arguments{
\item{v}{a vector}
}
\value{
rle but with values as vector value and lengths as vector name
}
\description{
get rle as a named vector
}
