% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_api.R
\name{safely_get_geocode}
\alias{safely_get_geocode}
\title{get geocode safely, it will never be interrupted}
\usage{
safely_get_geocode(address, apikey)
}
\arguments{
\item{address}{address, vector length of one}

\item{apikey}{google's apikey}
}
\value{
a list of google's geocode service result
}
\description{
get geocode safely, it will never be interrupted
}
